Fixed IPv6 tunneling issue
